+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v49] - unreleased
+
+### Added
+
+- J.S. Bach: Partita 2 BWV826
+
 ## [v48] - 2024-02-24
 
 ### Added